Generalize executables
