+++ pyproject.toml
@@ -34,6 +34,7 @@ importlib-metadata = {version = "^6.0", python = "<3.8"}
 commandlib = "^0.3.5"
 omegaconf = "^2.3.0"
 dill = "^0.3.6"
+typing-extensions = {version = "^4.5.0", python = "<3.11"}
 
 
 [tool.poetry.dev-dependencies]
